# openapi/api.yaml
openapi: 3.0.3
info:
  title: Codex Finance API
  version: 0.2.0
servers:
  - url: https://{host}/v1
    variables:
      host:
        default: api.example.com
components:
  securitySchemes:
    CognitoUserPool:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        txnId:
          type: string
        accountId:
          type: string
        categoryId:
          type: string
        type:
          type: string
          enum: [expense, income]
        amount:
          type: number
        currency:
          type: string
        fx_rate_to_base:
          type: number
        amount_base:
          type: number
        note:
          nullable: true
        tags:
          type: array
          items:
            type: string
        occurredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [txnId, accountId, categoryId, type, amount, currency, occurredAt]
    Budget:
      type: object
      properties:
        month:
          type: string
          pattern: "^\\d{4}-(0[1-9]|1[0-2])$"
        categoryId:
          nullable: true
        currency:
          type: string
        limit:
          type: number
        alertThreshold:
          type: number
        rollover:
          type: boolean
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [month, currency, limit]
paths:
...
  /transactions/{txnId}:
    parameters:
      - in: path
        name: txnId
        required: true
        schema:
          type: string
    get:
      summary: Fetch a transaction
      security:
        - CognitoUserPool: []
      responses:
        '200':
          description: Transaction detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404': { description: Not found }
    put:
      summary: Update a transaction
      security:
        - CognitoUserPool: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accountId: { type: string }
                categoryId: { type: string }
                type: { type: string, enum: [expense, income] }
                amount: { type: number }
                occurredAt: { type: string, format: date-time }
                fx_rate_to_base: { type: number }
                note: { type: string, nullable: true }
                tags:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
    delete:
      summary: Delete transaction
      security:
        - CognitoUserPool: []
      responses:
        '204': { description: Deleted }
...
  /budgets/{month}:
    parameters:
      - in: path
        name: month
        required: true
        schema:
          type: string
          pattern: "^\\d{4}-(0[1-9]|1[0-2])$"
    get:
      summary: List budgets for a given month
      security:
        - CognitoUserPool: []
      responses:
        '200':
          description: Budgets for the month
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Budget'
  /budgets/{month}/{categoryId}:
    parameters:
      - in: path
        name: month
        required: true
        schema:
          type: string
          pattern: "^\\d{4}-(0[1-9]|1[0-2])$"
      - in: path
        name: categoryId
        required: true
        schema:
          type: string
          description: Use "all" for an overall budget with no category
    get:
      summary: Fetch a specific budget entry
      security:
        - CognitoUserPool: []
      responses:
        '200':
          description: Budget entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
        '404': { description: Not found }
    put:
      summary: Update a budget entry
      security:
        - CognitoUserPool: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currency: { type: string }
                limit: { type: number }
                alertThreshold: { type: number }
                rollover: { type: boolean }
      responses:
        '200':
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
    delete:
      summary: Delete a budget entry
      security:
        - CognitoUserPool: []
      responses:
        '204': { description: Deleted }
...
security:
  - CognitoUserPool: []